+++ plugins/fixveins.cpp
@@ -56,8 +56,8 @@ command_result df_fixveins (color_ostream &out, vector <string> & parameters)
     int mineral_removed = 0, feature_removed = 0;
     int mineral_added = 0, feature_added = 0;
 
-    int num_blocks = 0, blocks_total = world->map.map_blocks.size();
-    for (int i = 0; i < blocks_total; i++)
+    size_t num_blocks = 0, blocks_total = world->map.map_blocks.size();
+    for (size_t i = 0; i < blocks_total; i++)
     {
         df::map_block *block = world->map.map_blocks[i];
         uint16_t has_mineral[16] = {0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0};
